+++ apps/myaccount/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @wso2is/myaccount
 
+## 2.7.50
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.organizations.v1@2.20.51
+  - @wso2is/common.branding.v1@2.20.51
+
 ## 2.7.49
 
 ### Patch Changes