+++ features/admin.email-management.v1/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @wso2is/admin.email-management.v1
 
+## 2.20.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.core.v1@2.21.21
+  - @wso2is/admin.branding.v1@2.20.51
+  - @wso2is/common.branding.v1@2.20.51
+
 ## 2.20.50
 
 ### Patch Changes