+++ features/admin.secrets.v1/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @wso2is/admin.secrets.v1
 
+## 2.20.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wso2is/admin.core.v1@2.21.21
+
 ## 2.20.50
 
 ### Patch Changes